# Optimization for getting value from  map, `arrayElement`(1/2)