--- a/src/Functions/array/arrayElement.cpp
+++ b/src/Functions/array/arrayElement.cpp
@@ -766,7 +766,7 @@ struct MatcherString
     const DataColumn & data;
     const IndexColumn & index;
 
-    inline bool match(size_t row_data, size_t row_index) const
+    bool match(size_t row_data, size_t row_index) const
     {
         auto data_ref = data.getDataAt(row_data);
         auto index_ref = index.getDataAt(row_index);
@@ -780,7 +780,7 @@ struct MatcherStringConst
     const DataColumn & data;
     const String & index;
 
-    inline bool match(size_t row_data, size_t /* row_index */) const
+    bool match(size_t row_data, size_t /* row_index */) const
     {
         auto data_ref = data.getDataAt(row_data);
         return index.size() == data_ref.size && memcmp(index.data(), data_ref.data, data_ref.size) == 0;
@@ -793,7 +793,7 @@ struct MatcherNumber
     const PaddedPODArray<DataType> & data;
     const PaddedPODArray<IndexType> & index;
 
-    inline bool match(size_t row_data, size_t row_index) const
+    bool match(size_t row_data, size_t row_index) const
     {
         return data[row_data] == static_cast<DataType>(index[row_index]);
     }
@@ -818,31 +818,35 @@ void FunctionArrayElement::executeMatchKeyToIndex(
     const Offsets & offsets, PaddedPODArray<UInt64> & matched_idxs, const Matcher & matcher)
 {
     size_t rows = offsets.size();
-    size_t matched_pos = 0;
+    size_t expected_match_pos = 0;
     bool matched = false;
-    for (size_t j = 0, end = offsets[0]; j < end; ++j)
+    if (rows)
     {
-        if (matcher.match(j, 0))
+        for (size_t j = offsets[-1], end = offsets[0]; j < end; ++j)
         {
-            matched_idxs.push_back(j + 1);
-            matched = true;
-            matched_pos = end + j;
-            break;
+            if (matcher.match(j, 0))
+            {
+                matched_idxs.push_back(j -  offsets[-1] + 1);
+                matched = true;
+                expected_match_pos = end + j - offsets[-1];
+                break;
+            }
         }
+        if (!matched)
+            matched_idxs.push_back(0);
     }
-    if (!matched)
-        matched_idxs.push_back(0);
-    /// In pratice, map keys are usually in the same order, it is worth a trying to
-    /// predicate the next key position as previous. So it can save a lot of comparisions.
+
+    /// In practice, map keys are usually in the same order, it is worth a try to
+    /// predicate the next key position as previous. So it can save a lot of comparisons.
     size_t i = 1;
     for (; i < rows; ++i)
     {
         const auto & begin = offsets[i - 1];
         const auto & end = offsets[i];
-        if (matched_pos < end && matcher.match(matched_pos, i))
+        if (expected_match_pos < end && matcher.match(expected_match_pos, i))
         {
-            matched_idxs.push_back(matched_pos - begin + 1);
-            matched_pos = end + matched_pos - begin;
+            matched_idxs.push_back(expected_match_pos - begin + 1);
+            expected_match_pos = end + expected_match_pos - begin;
         }
         else
             break;
@@ -850,8 +854,8 @@ void FunctionArrayElement::executeMatchKeyToIndex(
     for (; i < rows; ++i)
     {
         matched = false;
-        auto begin = offsets[i - 1];
-        auto end = offsets[i];
+        const auto & begin = offsets[i - 1];
+        const auto & end = offsets[i];
         for (size_t j = begin; j < end; ++j)
         {
             if (matcher.match(j, i))
@@ -861,6 +865,7 @@ void FunctionArrayElement::executeMatchKeyToIndex(
                 break;
             }
         }
+
         if (!matched)
             matched_idxs.push_back(0);
     }