<!---
A technical comment, you are free to remove or leave it as it is when PR is created
The following categories are used in the next scripts, update them accordingly
utils/changelog/changelog.py
tests/ci/cancel_and_rerun_workflow_lambda/app.py
-->
### Changelog category (leave one):
- Performance Improvement



### Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
...
Getting values from a map is widely used. In practice, the key structrues are usally the same in the same map column, we could try to predict the next row's key position and reduce the comparisons.

Some results are given as following
```sql
insert into test_table_map select CAST((['xyz', 'abc', '123'], [cast(number % 10 as String), 'dsdfsd', '123']), 'Map(String, String)') from numbers(20000000);

select a['xyz'] from test_table_map Format Null settings max_threads=1;
--- before
0 rows in set. Elapsed: 0.321 sec. Processed 20.00 million rows, 1.62 GB (62.25 million rows/s., 5.04 GB/s.)
0 rows in set. Elapsed: 0.322 sec. Processed 20.00 million rows, 1.62 GB (62.09 million rows/s., 5.03 GB/s.)
0 rows in set. Elapsed: 0.323 sec. Processed 20.00 million rows, 1.62 GB (61.98 million rows/s., 5.02 GB/s.)
--- after
0 rows in set. Elapsed: 0.316 sec. Processed 20.00 million rows, 1.62 GB (63.23 million rows/s., 5.12 GB/s.)
0 rows in set. Elapsed: 0.319 sec. Processed 20.00 million rows, 1.62 GB (62.70 million rows/s., 5.08 GB/s.)
0 rows in set. Elapsed: 0.316 sec. Processed 20.00 million rows, 1.62 GB (63.35 million rows/s., 5.13 GB/s.)


select a['123'] from test_table_map Format Null settings max_threads=1;
--- before
0 rows in set. Elapsed: 0.459 sec. Processed 20.00 million rows, 1.62 GB (43.56 million rows/s., 3.53 GB/s.)
0 rows in set. Elapsed: 0.460 sec. Processed 20.00 million rows, 1.62 GB (43.45 million rows/s., 3.52 GB/s.)
0 rows in set. Elapsed: 0.460 sec. Processed 20.00 million rows, 1.62 GB (43.44 million rows/s., 3.52 GB/s.)
--- after
0 rows in set. Elapsed: 0.318 sec. Processed 20.00 million rows, 1.62 GB (62.88 million rows/s., 5.09 GB/s.)
0 rows in set. Elapsed: 0.317 sec. Processed 20.00 million rows, 1.62 GB (63.12 million rows/s., 5.11 GB/s.)
0 rows in set. Elapsed: 0.317 sec. Processed 20.00 million rows, 1.62 GB (63.19 million rows/s., 5.12 GB/s.)
```

```sql
insert into test_table_map select if (number % 2 = 0, CAST((['xyz', 'abc', '123'], [cast(number % 10 as String), 'dsdfsd', '123']), 'Map(String, String)'), CAST((['123', 'abc', 'xyz'], ['123', 'dsdfsd', cast(number % 10 as String)]), 'Map(String, String)')) from numbers(20000000);

select a['xyz'] from test_table_map Format Null settings max_threads=1;
--- before
0 rows in set. Elapsed: 0.411 sec. Processed 20.00 million rows, 1.62 GB (48.66 million rows/s., 3.94 GB/s.)
0 rows in set. Elapsed: 0.422 sec. Processed 20.00 million rows, 1.62 GB (47.37 million rows/s., 3.84 GB/s.)
0 rows in set. Elapsed: 0.431 sec. Processed 20.00 million rows, 1.62 GB (46.46 million rows/s., 3.76 GB/s.)
--- after
0 rows in set. Elapsed: 0.408 sec. Processed 20.00 million rows, 1.62 GB (48.97 million rows/s., 3.97 GB/s.)
0 rows in set. Elapsed: 0.410 sec. Processed 20.00 million rows, 1.62 GB (48.79 million rows/s., 3.95 GB/s.)
0 rows in set. Elapsed: 0.414 sec. Processed 20.00 million rows, 1.62 GB (48.33 million rows/s., 3.91 GB/s.)
```

### Documentation entry for user-facing changes

- [ ] Documentation is written (mandatory for new features)

<!---
Directly edit documentation source files in the "docs" folder with the same pull-request as code changes

or

Add a user-readable short description of the changes that should be added to docs.clickhouse.com below.

At a minimum, the following information should be added (but add more as needed).
- Motivation: Why is this function, table engine, etc. useful to ClickHouse users?

- Parameters: If the feature being added takes arguments, options or is influenced by settings, please list them below with a brief explanation.

- Example use: A query or command.
-->


> Information about CI checks: https://clickhouse.com/docs/en/development/continuous-integration/