@antaljanosbenjamin 2023-10-27
@lgbo-ustc Generally we don't like to force push branches/merging master just to trigger a new run of CI runs. It takes resources without a good reason. Instead we prefer to check the test results and verify that the failures are not introduced by the PR.