@UnamedRus 2024-01-14
>  Getting values from a map is widely used. In practice, the key structrues are usally the same in the same map column, we could try to predict the next row's key position and reduce the comparisons.

About prediction, we probably can have more precise upper bound (it's can't exceed enum/dictionary index) for possible map key position for Enum & LowCardinality(Strings) types.